Extract character_db_ex from character.cpp
