+++ CMakeLists.txt
@@ -81,6 +81,7 @@ set(CORE_SOURCES
   src/crafting.cpp
   src/ctrl_file.cpp
   src/ctrl_inventory.cpp
+  src/db_character.cpp
   src/db_chara_chip.cpp
   src/db_item.cpp
   src/db_item_chip.cpp
